core: vm: sqlvm: parser: support IS TRUE / FALSE / UNKNOWN
To allow comparing directly with the value of a three-valued boolean
without being affected by the semantic of = operator, we have to support
SQL <boolean test> syntax.

To keep the parser simple, UNKNOWN is considered as a boolean literal
and it can be used outside of IS. It seems to me that the SQL standard
suggests that this is the correct behavior because UNKNOWN is listed
in both <boolean literal> and <truth value>.

This behavior seems to be different from what PostgreSQL and MySQL do.
They only recognize UNKNOWN when it is used at the right-hand side of IS
and consider it as an identifier when it is used outside of IS. However,
the standard says UNKNOWN is a reserved word, so automatically making
UNKNOWN an identifier looks non-standard to me.